pyproject.toml
[project]
name = "cocktaildispenser"
version = "0.1.0"
description = "A web and REST-based application for managing cocktail recipes and controlling automated liquid dispensing machines."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2.5",
    "djangorestframework>=3.16.1",
]
